Split up `wasm.rs` into own wasm module
