+++ p2panda-rs/src/wasm/entry.rs
@@ -1,0 +1,103 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+use std::convert::TryFrom;
+
+use serde::{Deserialize, Serialize};
+use wasm_bindgen::prelude::wasm_bindgen;
+use wasm_bindgen::JsValue;
+
+use crate::entry::{decode_entry as decode, sign_and_encode, Entry, EntrySigned, LogId, SeqNum};
+use crate::hash::Hash;
+use crate::operation::{Operation, OperationEncoded};
+
+use crate::wasm::error::jserr;
+use crate::wasm::KeyPair;
+
+/// Return value of [`sign_encode_entry`] that holds the encoded entry and its hash.
+#[derive(Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+struct SignEncodeEntryResult {
+    pub entry_encoded: String,
+    pub entry_hash: String,
+    pub operation_hash: String,
+}
+
+/// Returns a signed and encoded entry that can be published to a p2panda node.
+///
+/// `entry_backlink_hash`, `entry_skiplink_hash`, `seq_num` and `log_id` are obtained by querying
+/// the `getEntryArguments` method of a p2panda node.
+///
+/// `seq_num` and `log_id` are `i32` parameters even though they have 64 bits in the Bamboo spec.
+/// Webkit doesn't support `BigInt` so it can't handle those large values.
+#[wasm_bindgen(js_name = signEncodeEntry)]
+pub fn sign_encode_entry(
+    key_pair: &KeyPair,
+    encoded_operation: String,
+    entry_skiplink_hash: Option<String>,
+    entry_backlink_hash: Option<String>,
+    seq_num: i32,
+    log_id: i32,
+) -> Result<JsValue, JsValue> {
+    // If skiplink_hash exists construct Hash
+    let skiplink_hash = match entry_skiplink_hash {
+        Some(hash) => Some(jserr!(Hash::new(&hash))),
+        None => None,
+    };
+
+    // If backlink_hash exists construct Hash
+    let backlink_hash = match entry_backlink_hash {
+        Some(hash) => Some(jserr!(Hash::new(&hash))),
+        None => None,
+    };
+
+    // Create SeqNum instance
+    let seq_num = jserr!(SeqNum::new(seq_num.into()));
+
+    // Convert to Operation
+    let operation_encoded = jserr!(OperationEncoded::new(&encoded_operation));
+    let operation = jserr!(Operation::try_from(&operation_encoded));
+
+    // Create Entry instance
+    let entry = jserr!(Entry::new(
+        &LogId::new(log_id.into()),
+        Some(&operation),
+        skiplink_hash.as_ref(),
+        backlink_hash.as_ref(),
+        &seq_num,
+    ));
+
+    // Finally sign and encode entry
+    let entry_signed = jserr!(sign_and_encode(&entry, key_pair.as_inner()));
+
+    // Serialise result to JSON
+    let result = jserr!(wasm_bindgen::JsValue::from_serde(&SignEncodeEntryResult {
+        entry_encoded: entry_signed.as_str().into(),
+        entry_hash: entry_signed.hash().as_str().into(),
+        operation_hash: operation_encoded.hash().as_str().into(),
+    }));
+    Ok(result)
+}
+
+/// Decodes an entry and optional operation given their encoded form.
+#[wasm_bindgen(js_name = decodeEntry)]
+pub fn decode_entry(
+    entry_encoded: String,
+    operation_encoded: Option<String>,
+) -> Result<JsValue, JsValue> {
+    // Convert encoded operation
+    let operation_encoded = match operation_encoded {
+        Some(msg) => {
+            let inner = jserr!(OperationEncoded::new(&msg));
+            Some(inner)
+        }
+        None => None,
+    };
+
+    // Convert encoded entry
+    let entry_signed = jserr!(EntrySigned::new(&entry_encoded));
+    let entry: Entry = jserr!(decode(&entry_signed, operation_encoded.as_ref()));
+
+    // Serialise struct to JSON
+    let result = jserr!(wasm_bindgen::JsValue::from_serde(&entry));
+    Ok(result)
+}
